Pman/Roo.php

+++ Pman/Roo.php
@@ -689,8 +689,14 @@ class Pman_Roo extends Pman
         
         //echo '<PRE>';print_r($this->cols); exit;
         $options = &PEAR::getStaticProperty('DB_DataObject','options');
-         
-        echo '<PRE>';print_r($options["ini_{$x->_database}"] );exit;
+        $reader = $options["ini_{$x->_database}"] .'.reader';
+        if (!file_exists( $reader )) {
+            return;
+        }
+        
+        $rdata = unserialize(file_get_contents($reader));
+        
+        echo '<PRE>';print_r($rdata);exit;
         
         file_put_contents($options["ini_{$x->_database}"] . '.reader', serialize($out));
